Decompile func0f02dff0

+++ src/include/game/chr/chraction.h
@@ -3,7 +3,7 @@
 #include <ultra64.h>
 #include "types.h"
 
-f32 func0f02dff0(u32 animnum);
+f32 func0f02dff0(s16 animnum);
 u32 func0f02e064(struct chrdata *chr);
 bool weaponIsOneHanded(struct prop *prop);
 f32 chrGetRangedSpeed(struct chrdata *chr, f32 min, f32 max);
